// src/app/models/photo.ts
export class Photo {
  comments: Number;
  downloads: Number;
  favorites: Number;
  id: Number;
  imageHeight: Number;
  imageSize: Number;
  imageWidth: Number;
  largeImageURL: string;
  likes: Number;
  pageURL: string;
  previewHeight: Number;
  previewURL: string;
  previewWidth: Number;
  tags: string;
  type: string;
  user: string;
  userImageURL: string;
  user_id: Number;
  views: Number;
  webformatHeight: Number;
  webformatURL: string;
  webformatWidth: Number;
}